Fixing error in widgest manager
