+++ index.html
@@ -26,36 +26,7 @@
     <div class='workspace-container'>
         <div class='map-container' id='map-container'>
             <div class='map-view' id='map'></div>
-
-<!--            <div class='map-toolbar map-toolbar-bar' id='map-toolbar'>
-                <div class='map-zoom'>
-                    <div class='map-zoom-indicator'>z&#8202;<span class='map-zoom-quantity'></span></div>
-                    <button class='map-zoom-in' id='zoom-in' data-tooltip='Zoom in'><i class='btm bt-plus'></i></button>
-                    <button class='map-zoom-out' id='zoom-out' data-tooltip='Zoom out'><i class='btm bt-minus'></i></button>
-                </div>
-                <div class='map-location-bar'>
-                    <div class='map-search-icon' data-tooltip='Search for location'><i class='btm bt-search'></i></div>
-                    <input class='map-search-input' placeholder='' spellcheck='false'>
-                    <div class='map-latlng-label'></div>
-                    <div class='map-save-icon' data-tooltip='Bookmark location' data-tooltip-alignment='right'><i class='btm bt-star'></i></div>
-                    <div class='map-bookmarks'>
-                        <button class='map-bookmarks-button' id='bookmarks' data-tooltip='Bookmarks' data-tooltip-alignment='right'>
-                            <i class='btm bt-bookmark'></i>
-                        </button>
-                        <div class='map-bookmarks-menu'></div>
-                    </div>
-                    <div class='map-search-results'></div>
-                </div>
-                <div class='geolocator'>
-                    <button class='geolocator-button' id='geolocator' data-tooltip='Locate me' data-tooltip-alignment='right'>
-                        <i class='btm bt-map-arrow'></i>
-                        <i class='btm bt-sync bt-spin'></i>
-                    </button>
-                </div>
-                <div class='map-toolbar-toggle' id='map-toolbar-toggle' data-tooltip='Toogle map toolbar' data-tooltip-alignment='right'></div>
-            </div>-->
             <div class='map-loading' id='map-loading'></div>
-
             <div id='map-panel'></div>
         </div>
         <div class='divider' id='divider'><span class='divider-affordance'></span></div>
